requirements.txt
requests
python-dotenv
supabase
feedparser
